Improve Item Cursor Graphic enum format

+++ Source/qol/itemlabels.cpp
@@ -103,7 +103,7 @@ void AddItemToLabelQueue(int id, Point position)
 
 	int nameWidth = GetLineWidth(textOnGround);
 	nameWidth += MarginX * 2;
-	int index = ItemCAnimTbl[item._iCurs];
+	int index = ItemCAnimTbl[static_cast<uint8_t>(item._iCurs)];
 	if (!labelCenterOffsets[index]) {
 		std::pair<int, int> itemBounds = ClxMeasureSolidHorizontalBounds((*item.AnimInfo.sprites)[item.AnimInfo.currentFrame]);
 		labelCenterOffsets[index].emplace((itemBounds.first + itemBounds.second) / 2);
